Move measurement code to a module
